Create two tabs in left panel
